node installation
intersting solution - the installation of all the package is executed in the docker image.
I'm running the typescript watcher outside docker, in pycharm.
I tried running it inside (and that's probably possible as well ) but this way it works easier.
it works fine because of file mapping - the node_modules are installed inside docker - but available outside just as well.

so the typescript compiles the angular2 componenets outdside - but the js files are served as static files by django from inside the container.